server/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
anthropic
ffmpeg-python
openai-whisper
pydantic
aiofiles
